parallel: Move curses ui to own file
